Move every post screen to its folder
