+++ src/screens/index.js
@@ -1,5 +1,4 @@
 export { names, titles } from "./names";
 
 export { Home } from "./Home";
-export { Post } from "./Post";
 export { Search } from "./Search";
