+++ src/screens/post/index.js
@@ -1,0 +1,5 @@
+export { ArticuloPost } from "./ArticuloPost";
+export { Post } from "./Post";
+export { RegalarPost } from "./RegalarPost";
+export { StandarPost } from "./StandarPost";
+export { VenderPost } from "./VenderPost";
